Move test code for deprecated-local-action rule into the rule file
